src/com/tasktoys/java8ws/mikan/ch2/ex06/CharacterStream.java
/*
 * Copyright(C) 2014-2015 Java 8 Workshop participants. All rights reserved.
 * https://github.com/Java8Workshop/Exercises
 */
package com.tasktoys.java8ws.mikan.ch2.ex06;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.IntStream;
import java.util.stream.Stream;

/**
 *
 * @author mikan
 */
public class CharacterStream {

    private CharacterStream() {

    }

    public static Stream<Character> characterStream(String s) {
        List<Character> result = new ArrayList<>();
        for (char c : s.toCharArray()) {
            result.add(c);
        }
        return result.stream();
    }

    public static Stream<Character> characterStream2(String s) {
        return IntStream.range(0, s.length()).map(s::charAt).mapToObj(i -> (char) i);
    }
}
